Return match seconds as float instead of int

+++ game/shared/sdk/sdk_gamerules.h
@@ -476,7 +476,7 @@ public:
 
 	bool IsIntermissionState();
 	int GetShieldRadius(int team, bool isTaker);
-	int GetMatchDisplayTimeSeconds(bool addInjuryTime = true, bool getCountdownAtIntermissions = true);
+	float GetMatchDisplayTimeSeconds(bool addInjuryTime = true, bool getCountdownAtIntermissions = true);
 
 	float GetStrengthScalingCoeff();
